+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 7.8
 - Fix check students Course Status in PrintClassLists.php, PrintClassPictures.php, ClassSearchWidget.fnc.php, Referrals.php, EmailReferral.fnc.php & Widgets.fnc.php
 - Add Include Inactive Students checkbox in MasterScheduleReport.php & RequestsReport.php
 - Fix unset current student (check Course Status) when MP updated in Side.php
+- SQL fix Discipline Referrals using WHERE EXISTS in WidgetExtra.php
 
 Changes in 7.7
 --------------